Add single sentence to advice naming language in pull requests
The Year of Programming 2017 Challenges has attracted a lot of
participants. This translates to quite an immense number of pull
requests as these participants complete and submit their solutions.

Following the requirement stated in the slack announcement of a
minimum of three reviews per pull request for merging, it would
be helpful for reviewers if pull requests stated both challenge no.
as well as programming languge.

This is as some participants may know only and thus only review a
limited set of programming languages.

It would also help with organisation of pull requests/solution
submissions.